requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
choreographer==1.2.1
google-auth==2.43.0
google-genai==1.53.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
kaleido==1.2.0
logistro==2.0.1
narwhals==2.13.0
orjson==3.11.4
packaging==25.0
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-timeout==2.4.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
simplejson==3.20.2
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
